PR comment rsolution: rename Pgn to PGN and PgnConf to PGNConf
