+++ go/pkg/pgn/config/config_test.go
@@ -46,5 +46,5 @@ func InitTestConfig(cfg *config.Config) {
 func CheckTestConfig(t *testing.T, cfg *config.Config, id string) {
 	envtest.CheckTest(t, nil, &cfg.Metrics, nil, &cfg.SD, id)
 	logtest.CheckTestLogging(t, &cfg.Logging, id)
-	configtest.CheckTestPGN(t, &cfg.Pgn, id)
+	configtest.CheckTestPGN(t, &cfg.PGN, id)
 }
